backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django project for an interview"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.1",
]